+++ R/utils_assert.R
@@ -287,6 +287,22 @@ assert_script <- function() {
     "that would source _targets.R again and cause infinite recursion."
   )
   assert_not_in(vars, choices, msg)
+  msg <- paste(
+    "Do not use %s() from {devtools} or {pkgload} to load ",
+    "packages or custom functions/globals for {targets}. If you do, ",
+    "custom functions will go to a package environment where {targets} ",
+    "may not track them, and the loaded data will not be available in ",
+    "parallel workers created by tar_make_clustermq() or tar_make_future(). ",
+    "Read https://books.ropensci.org/targets/practices.html#loading-and-configuring-r-packages ", # nolint
+    "and https://books.ropensci.org/targets/practices.html#packages-based-invalidation ", # nolint
+    "for the correct way to load and configure packages for {targets} pipelines. ",
+    "Suppress this warning with Sys.getenv(TAR_WARN = \"false\")."
+  )
+  for (loader in c("load_all", "load_code", "load_data", "load_dll")) {
+    if (Sys.getenv("TAR_WARN") != "false" && loader %in% vars) {
+      warn_validate(sprintf(msg, loader))
+    }
+  }
 }
 
 assert_true <- function(condition, msg = NULL) {