+++ tests/testthat/test-class_imports.R
@@ -67,6 +67,9 @@ tar_test("imports_validate()", {
 })
 
 tar_test("imports setting works", {
+  old <- Sys.getenv("TAR_WARN")
+  on.exit(Sys.setenv(TAR_WARN = old))
+  Sys.setenv(TAR_WARN = "false") # This test has a good reason for load_all().
   skip_if_not_installed("pkgload")
   dir_create("pkg")
   dir_create(file.path("pkg", "R"))